util: don't cast potential signed 64-bit type to int for error checking
If times() returns a value long which ends up being negative when cast
to an integer, then we erroneously fail in checkedErrno even though
the return is valid:

[    1.000] caught exception, terminating: loop times: failed with -16163 err=Unknown error 16163(16163)
[    1.000] caught exception, terminating: loop times: failed with -16163 err=Unknown error 16163(16163)
[    1.000] uncaught thread exception: loop times: failed with -16163 err=Unknown error 16163(16163)

Fix this by ensuring we don't cast a signed 64-bit to 32-bit.

+++ util.h
@@ -57,9 +57,9 @@ inline std::string leftpad(std::string x, size_t n) {
 struct InterruptedException : std::exception {};
 
 template <typename... T>
-inline int checkedErrno(int res, const T&... vals) {
+inline int checkedErrno(int64_t res, const T&... vals) {
   if (res < 0) {
-    int error = res == -1 ? errno : -res;
+    int64_t error = res == -1 ? errno : -res;
     if (error == EINTR) {
       throw InterruptedException{};
     }
